requirements.txt
amqp
asgiref
billiard
celery
certifi
charset-normalizer
click
colorama
Django
django-admin
django-cors-headers
django-environ
django-extensions
djangorestframework
djangorestframework_simplejwt
dnspython
drf-yasg
idna
kombu
mysqlclient
packaging
pillow
PyJWT
pytz
redis
requests
six
sqlparse
tzdata
urllib3
vine
requests
dj_database_url
redis
amqp
gunicorn
psycopg2-binary
django-celery-results
whitenoise